requirements.txt
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764727380/work
certifi==2022.12.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179605388/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
click==8.1.3
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1679811337757/work
diffusers==0.10.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1680834529360/work
Flask==2.2.3
frozenlist==1.3.3
ftfy==6.1.1
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1666808701622/work
huggingface-hub==0.13.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1679914882579/work
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1674135922531/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict==6.0.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1675642633393/work
openai==0.27.4
opencv-python==4.7.0.72
packaging==23.0
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1675487390302/work
psutil==5.9.4
PyAudio==0.2.13
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
PyYAML==6.0
regex==2023.3.23
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1680286922386/work
scipy==1.10.1
SpeechRecognition==3.10.0
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1679342590084/work
tokenizers==0.13.3
torch==2.1.0.dev20230410
torchaudio==2.1.0.dev20230410
torchvision==0.16.0.dev20230410
tqdm==4.65.0
transformers==4.27.4
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
